RelationShips

+++ database/migrations/2019_07_07_195330_create_questions_table.php
@@ -21,8 +21,8 @@ class CreateQuestionsTable extends Migration
             $table->text('body');
 
             
-            $table->integer('category_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             
             $table->timestamps();
         });
